Move lib/ to src/
