+++ package.json
@@ -22,13 +22,13 @@
     "coveralls": "~2.11.2"
   },
   "scripts": {
-    "grammar": "node_modules/.bin/pegjs --cache lib/grammar.pegjs",
+    "grammar": "node_modules/.bin/pegjs --cache src/grammar.pegjs",
     "test": "node_modules/.bin/mocha",
     "test-spec": "node_modules/.bin/mocha test/spec",
     "gen-spec": "node share/gen-spec.js",
     "coverage": "node_modules/.bin/mocha -R html-cov test test/spec > coverage.html",
     "report-coverage": "node_modules/.bin/mocha -R mocha-lcov-reporter test test/spec | ./node_modules/.bin/coveralls",
-    "browser": "node_modules/.bin/browserify lib/browser.js --outfile dist/browser.js --standalone hummingbird"
+    "browser": "node_modules/.bin/browserify src/browser.js --outfile dist/browser.js --standalone hummingbird"
   }
 }
