+++ src/browser.js
