+++ src/parser-extension.js
