+++ src/spec-parser.js
