+++ src/targets/javascript.js
