+++ src/targets/llvm/native-function.js
