+++ src/targets/llvm/native-object.js
