+++ src/targets/llvm/util.js
