+++ src/typesystem/scope.js
